linting test config
