+++ tests_config/raw-serializer.js
@@ -1,6 +1,4 @@
 // source: https://github.com/prettier/prettier/blob/ee2839bacbf6a52d004fa2f0373b732f6f191ccc/tests_config/raw-serializer.js
-'use strict';
-
 const RAW = Symbol.for('raw');
 
 module.exports = {
