+++ tests_config/run_spec.js
@@ -1,16 +1,19 @@
 // source: https://github.com/prettier/prettier/blob/ee2839bacbf6a52d004fa2f0373b732f6f191ccc/tests_config/run_spec.js
-'use strict';
-
 const fs = require('fs');
 const path = require('path');
+const {
+  read,
+  mergeDefaultOptions,
+  prettyprint,
+  raw,
+  parse
+} = require('./helpers');
 
-const AST_COMPARE = process.env['AST_COMPARE'];
+const { AST_COMPARE } = process.env;
 
-const prettier = require('prettier');
-
-function run_spec(dirname, options) {
-  fs.readdirSync(dirname).forEach(filename => {
-    const filepath = dirname + '/' + filename;
+function runSpec(dirname, options) {
+  fs.readdirSync(dirname).forEach((filename) => {
+    const filepath = `${dirname}/${filename}`;
     if (
       path.extname(filename) !== '.snap' &&
       fs.lstatSync(filepath).isFile() &&
@@ -45,16 +48,16 @@ function run_spec(dirname, options) {
       const output = prettyprint(input, mergedOptions);
       test(filename, () => {
         expect(
-          raw(source + '~'.repeat(mergedOptions.printWidth) + '\n' + output)
+          raw(`${source + '~'.repeat(mergedOptions.printWidth)}\n${output}`)
         ).toMatchSnapshot();
       });
 
       if (AST_COMPARE) {
         test(`${filepath} parse`, () => {
-          const compareOptions = Object.assign({}, mergedOptions);
+          const compareOptions = { ...mergedOptions };
           delete compareOptions.cursorOffset;
           const astMassaged = parse(input, compareOptions);
-          let ppastMassaged = undefined;
+          let ppastMassaged;
 
           expect(() => {
             ppastMassaged = parse(
@@ -73,45 +76,4 @@ function run_spec(dirname, options) {
   });
 }
 
-global.run_spec = run_spec;
-
-function parse(string, opts) {
-  return prettier.__debug.parse(string, opts, /* massage */ true).ast;
-}
-
-function prettyprint(src, options) {
-  const result = prettier.formatWithCursor(src, options);
-  if (options.cursorOffset >= 0) {
-    result.formatted =
-      result.formatted.slice(0, result.cursorOffset) +
-      '<|>' +
-      result.formatted.slice(result.cursorOffset);
-  }
-  return result.formatted;
-}
-
-function read(filename) {
-  return fs.readFileSync(filename, 'utf8');
-}
-
-/**
- * Wraps a string in a marker object that is used by `./raw-serializer.js` to
- * directly print that string in a snapshot without escaping all double quotes.
- * Backticks will still be escaped.
- */
-function raw(string) {
-  if (typeof string !== 'string') {
-    throw new Error('Raw snapshots have to be strings.');
-  }
-  return { [Symbol.for('raw')]: string };
-}
-
-function mergeDefaultOptions(parserConfig) {
-  return Object.assign(
-    {
-      plugins: [path.dirname(__dirname)],
-      printWidth: 80
-    },
-    parserConfig
-  );
-}
+global.run_spec = runSpec;
